chore: rename modules/folders for rv32im
