+++ vm/src/arch/chips.rs
@@ -32,14 +32,7 @@ use crate::{
     old::{
         alu::ArithmeticLogicChip, shift::ShiftChip, uint_multiplication::UintMultiplicationChip,
     },
-    rv32im::{
-        base_alu::Rv32BaseAluChip, branch_eq::Rv32BranchEqualChip,
-        branch_lt::Rv32BranchLessThanChip, divrem::Rv32DivRemChip, less_than::Rv32LessThanChip,
-        load_sign_extend::Rv32LoadSignExtendChip, loadstore::Rv32LoadStoreChip,
-        mul::Rv32MultiplicationChip, mulh::Rv32MulHChip, rv32_auipc::Rv32AuipcChip,
-        rv32_hintstore::Rv32HintStoreChip, rv32_jal_lui::Rv32JalLuiChip, rv32_jalr::Rv32JalrChip,
-        shift::Rv32ShiftChip,
-    },
+    rv32im::*,
     system::program::ExecutionError,
 };
