+++ vm/src/rv32im/adapters/mul.rs
