+++ vm/src/rv32im/adapters/vec_heap.rs
