+++ vm/src/rv32im/hintstore/core.rs
