+++ vm/src/rv32im/hintstore/tests.rs
@@ -27,7 +27,7 @@ use crate::{
     },
     rv32im::{
         adapters::{compose, Rv32HintStoreAdapterChip, RV32_CELL_BITS, RV32_REGISTER_NUM_LIMBS},
-        rv32_hintstore::Rv32HintStoreCoreCols,
+        hintstore::Rv32HintStoreCoreCols,
     },
     system::vm::{chip_set::BYTE_XOR_BUS, Streams},
     utils::{u32_into_limbs, u32_sign_extend},
