+++ vm/src/rv32im/mod.rs
@@ -1,15 +1,31 @@
 pub mod adapters;
-pub mod base_alu;
-pub mod branch_eq;
-pub mod branch_lt;
-pub mod divrem;
-pub mod less_than;
-pub mod load_sign_extend;
-pub mod loadstore;
-pub mod mul;
-pub mod mulh;
-pub mod rv32_auipc;
-pub mod rv32_hintstore;
-pub mod rv32_jal_lui;
-pub mod rv32_jalr;
-pub mod shift;
+
+mod auipc;
+mod base_alu;
+mod branch_eq;
+mod branch_lt;
+mod divrem;
+mod hintstore;
+mod jal_lui;
+mod jalr;
+mod less_than;
+mod load_sign_extend;
+mod loadstore;
+mod mul;
+mod mulh;
+mod shift;
+
+pub use auipc::*;
+pub use base_alu::*;
+pub use branch_eq::*;
+pub use branch_lt::*;
+pub use divrem::*;
+pub use hintstore::*;
+pub use jal_lui::*;
+pub use jalr::*;
+pub use less_than::*;
+pub use load_sign_extend::*;
+pub use loadstore::*;
+pub use mul::*;
+pub use mulh::*;
+pub use shift::*;
