+++ vm/src/system/vm/chip_set.rs
@@ -66,20 +66,7 @@ use crate::{
             Rv32HintStoreAdapterChip, Rv32JalrAdapterChip, Rv32LoadStoreAdapterChip,
             Rv32MultAdapterChip, Rv32RdWriteAdapterChip, Rv32VecHeapAdapterChip,
         },
-        base_alu::{BaseAluCoreChip, Rv32BaseAluChip},
-        branch_eq::{BranchEqualCoreChip, Rv32BranchEqualChip},
-        branch_lt::{BranchLessThanCoreChip, Rv32BranchLessThanChip},
-        divrem::{DivRemCoreChip, Rv32DivRemChip},
-        less_than::{LessThanCoreChip, Rv32LessThanChip},
-        load_sign_extend::{LoadSignExtendCoreChip, Rv32LoadSignExtendChip},
-        loadstore::{LoadStoreCoreChip, Rv32LoadStoreChip},
-        mul::{MultiplicationCoreChip, Rv32MultiplicationChip},
-        mulh::{MulHCoreChip, Rv32MulHChip},
-        rv32_auipc::{Rv32AuipcChip, Rv32AuipcCoreChip},
-        rv32_hintstore::{Rv32HintStoreChip, Rv32HintStoreCoreChip},
-        rv32_jal_lui::{Rv32JalLuiChip, Rv32JalLuiCoreChip},
-        rv32_jalr::{Rv32JalrChip, Rv32JalrCoreChip},
-        shift::{Rv32ShiftChip, ShiftCoreChip},
+        *,
     },
     system::{
         connector::VmConnectorChip,
